Crate solana_rpc

Modules

cluster_tpu_info 🔒

entry_notifier_interface

max_slots

optimistically_confirmed_bank_tracker

The optimistically_confirmed_bank_tracker module implements a threaded service to track the most recent optimistically confirmed bank for use in rpc services, and triggers gossip subscription notifications

parsed_token_accounts

rpc

The rpc module implements the Solana RPC interface.

rpc_cache 🔒

rpc_completed_slots_service

rpc_health

rpc_pubsub

The pubsub module implements a threaded subscription service on client RPC request

rpc_pubsub_service

The pubsub module implements a threaded subscription service on client RPC request

rpc_service

The rpc_service module implements the Solana JSON RPC service.

rpc_subscription_tracker

rpc_subscriptions

The pubsub module implements a threaded subscription service on client RPC request

transaction_notifier_interface

transaction_status_service