Module solana_rpc::rpc

The rpc module implements the Solana RPC interface.

Modules

rpc_accounts

rpc_accounts_scan

rpc_bank

rpc_deprecated_v1_7

rpc_deprecated_v1_9

rpc_full

rpc_minimal

rpc_obsolete_v1_7

Structs

JsonRpcConfig

JsonRpcRequestProcessor

RpcBigtableConfig

Constants

MAX_BASE58_SIZE 🔒

MAX_BASE64_SIZE 🔒

MAX_REQUEST_BODY_SIZE

PERFORMANCE_SAMPLES_LIMIT

Functions

_send_transaction 🔒

check_is_at_least_confirmed 🔒

create_test_transaction_entries

create_validator_exit

decode_and_deserialize 🔒

encode_account 🔒

get_encoded_account 🔒

get_spl_token_mint_filter 🔒

Analyze custom filters to determine if the result will be a subset of spl-token accounts by mint. NOTE: optimize_filters() should almost always be called before using this method because of the strict match on MemcmpEncodedBytes::Bytes.

get_spl_token_owner_filter 🔒

Analyze custom filters to determine if the result will be a subset of spl-token accounts by owner. NOTE: optimize_filters() should almost always be called before using this method because of the strict match on MemcmpEncodedBytes::Bytes.

get_token_program_id_and_mint 🔒

Analyze a passed Pubkey that may be a Token program id or Mint address to determine the program id and optional Mint

is_finalized 🔒

new_response 🔒

optimize_filters 🔒

populate_blockstore_for_tests

rpc_perf_sample_from_perf_sample 🔒

sanitize_transaction 🔒

verify_and_parse_signatures_for_address_params 🔒

verify_filter 🔒

verify_hash 🔒

verify_pubkey

verify_signature 🔒

verify_token_account_filter 🔒

verify_transaction 🔒

Type Definitions

RpcCustomResult 🔒